Implemented projector in java, and have its own prediction (instead of using lda prediction)
Added Jama library for matrix opertations.

Fixed bugs in run-batch, and the python codes for parsing output. Now can find the json file at the right location, and the columns and colors are displayed correctly.

Notice: to run synthetic data, in test set -Dplusone.useNewJsonFormat to false, and true if real data.